+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 LogicNG uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.4.2] - 2023-xx-xx
+## [2.5.0] - 2023-xx-xx
+
+### Added
+
+- Class `UBTree` offers new method `generateSubsumedUBTree` to directly generate a subsumed UBTree for the given sets. 
 
 ### Changed
 
+- UBTree data structure now supports empty sets.
 - Added side effect note in `SATSolver` for the four assumption solving methods. 
 
 ### Fixed